moveout store to each domain
