+++ frontend/src/domain/common.ts
@@ -1,0 +1,4 @@
+// All database model's id is int64
+type Id = number
+
+export { type Id }
