+++ frontend/src/lib/todo-calendar.tsx
@@ -1,5 +1,5 @@
 import * as React from "react"
-import { type DateKey, toDateKey } from "./state"
+import { type DateKey, toDateKey } from "#/domain/todo/store"
 
 const CurrentDateContext = React.createContext<{
   currentDate: Date
