+++ frontend/src/routes/-components/create-todo/create-todo.tsx
@@ -5,7 +5,9 @@ import { CalendarIcon, Check, ChevronsUpDown } from "lucide-react"
 import { toast } from "sonner"
 import { useScrollLock } from "#/lib/scroll-lock"
 import { cn } from "#/lib/utils"
-import { toDateKey, useBowlContext, useFlakeContext, useTodoContext } from "#/lib/state"
+import { useBowlContext } from "#/domain/bowl/store"
+import { useFlakeContext } from "#/domain/flake/store"
+import { useTodoContext, toDateKey } from "#/domain/todo/store"
 import { Popover, PopoverContent, PopoverTrigger } from "#/components/ui/popover"
 import { Button } from "#/components/ui/button"
 import { Command, CommandInput, CommandItem, CommandList } from "#/components/ui/command"
